requirements.txt
psycopg2-binary
requests
pandas
openpyxl
boto3
python-dotenv
